bridge package moved to cqldriver
